+++ src/components/loginButton/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const LoginButton = () => (
+  <>
+    <button className='create-account'>Create Account</button>
+    <button className="login">Login</button>
+  </>
+);
+
+export default LoginButton;
