+++ src/components/loginButton/styles.css
@@ -1,0 +1,3 @@
+.login ,.create-account{
+    width:30%;
+}
