+++ src/pages/login.css
@@ -1,52 +1,47 @@
 .main-login-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-content: center;
-    background-color: #fff;
-    margin-top: 200px;
-    padding: 50px;
-    width: 50%;
-    margin-left: 480px;
-  }
-  
-  input {
-    min-width: 80%;
-    padding: 10px;
-    margin: 8px 0;
-    border: 1px solid #ccc;
-    font-size: 14px;
-    transition: transform 0.2s ease, box-shadow 0.2s ease;
-    margin: 15px;
-    color: #7b2cbf;
-    border: none;
-  }
-  
-  .details-container {
-    display: inline;
-  }
-  
-  .login-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .sign-up {
-    width: 30%;
-  }
-  
-  .button-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    gap: 80px;
-    margin-top: 25px;
-  }
-  
-  .head {
-    font-family: 'bold';
-    font-size: 30px;
-  }
-  
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-content: center;
+  background-color: #fff;
+  margin-top: 200px;
+  padding: 50px;
+  width: 50%;
+  margin-left: 480px;
+}
+
+input {
+  min-width: 80%;
+  padding: 10px;
+  margin: 8px 0;
+  border: 1px solid #ccc;
+  font-size: 14px;
+  transition: transform 0.2s ease, box-shadow 0.2s ease;
+  margin: 15px;
+  color: #7b2cbf;
+  border: none;
+}
+
+.details-container {
+  display: inline;
+}
+
+.login-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.button-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  gap: 80px;
+  margin-top: 25px;
+}
+
+.welcome-title {
+  font-family: 'bold';
+  font-size: 30px;
+}
